Split mdclient into remote client and job definition sections.
"mdclient.py" contained utilities to connect to a Pyro queuing server,
and the class Mdjob which defines a job. This commit splits these
orthogonal activities. The Mdjob class is brought back to mdrunr, and
the utilities to connect to a queuing server are moved to a new
"client.py" file.