Added name.{out,err} files for jobs.
The Mdjob class now automatically sends the job output to the file
'job_name.out' and the error to 'job_name.err'. The 'combine_err_out'
boolean keyword argument allows stdout and stderr to be combined.

Also added a 'config.py' file for system-wide constants, like the name
of the output file and the standard error file.