Better Queue data structure.
This commit provides a subclass of Queue.Queue that allows (hopefully)
threadsafe modification of objects within the queue, rather than just
pushing and popping objects onto and off the queue. There should be
enough documentation in the module to make the new data structure
understandable.

Having this should hopefully simplify the implementation of Mdrunr.

I have also included a few unittests to check stuff works more or less
as expected.